+++ core/lib/appMain.d.ts
@@ -4,6 +4,6 @@ export declare class appMain {
     private readonly bw;
     private constructor();
     openDevTools(): void;
-    static initRenderer(path_htm: string, o: object): BrowserWindow;
+    static initRenderer(path_htm: string, version: string, o: object): BrowserWindow;
 }
 //# sourceMappingURL=appMain.d.ts.map